prevent subflows from streaming and other code cleanup
