+++ packages/components/nodes/agents/AirtableAgent/AirtableAgent.ts
@@ -107,7 +107,6 @@ class Airtable_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
