+++ packages/components/nodes/agents/AutoGPT/AutoGPT.ts
@@ -116,7 +116,6 @@ class AutoGPT_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
