+++ packages/components/nodes/agents/BabyAGI/BabyAGI.ts
@@ -76,7 +76,6 @@ class BabyAGI_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
