+++ packages/components/nodes/agents/CSVAgent/CSVAgent.ts
@@ -93,7 +93,6 @@ class CSV_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
