+++ packages/components/nodes/agents/ReActAgentChat/ReActAgentChat.ts
@@ -91,7 +91,6 @@ class ReActAgentChat_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
