+++ packages/components/nodes/agents/ReActAgentLLM/ReActAgentLLM.ts
@@ -80,7 +80,6 @@ class ReActAgentLLM_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
