+++ packages/components/nodes/agents/XMLAgent/XMLAgent.ts
@@ -134,7 +134,6 @@ class XMLAgent_Agents implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                // streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
