+++ packages/components/nodes/chains/ApiChain/OpenAPIChain.ts
@@ -87,7 +87,6 @@ class OpenApiChain_Chains implements INode {
                 if (options.shouldStreamResponse) {
                     streamResponse(sseStreamer, chatId, e.message)
                 }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
