+++ packages/components/nodes/chains/ConversationChain/ConversationChain.ts
@@ -128,7 +128,6 @@ class ConversationChain_Chains implements INode {
                 if (options.shouldStreamResponse) {
                     streamResponse(options.sseStreamer, options.chatId, e.message)
                 }
-                // streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
