+++ packages/components/nodes/chains/ConversationalRetrievalQAChain/ConversationalRetrievalQAChain.ts
@@ -213,7 +213,6 @@ class ConversationalRetrievalQAChain_Chains implements INode {
                 if (options.shouldStreamResponse) {
                     streamResponse(options.sseStreamer, options.chatId, e.message)
                 }
-                // streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
