+++ packages/components/nodes/chains/MultiRetrievalQAChain/MultiRetrievalQAChain.ts
@@ -97,7 +97,6 @@ class MultiRetrievalQAChain_Chains implements INode {
                 if (options.shouldStreamResponse) {
                     streamResponse(options.sseStreamer, options.chatId, e.message)
                 }
-                // streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
