+++ packages/components/nodes/chains/RetrievalQAChain/RetrievalQAChain.ts
@@ -74,7 +74,6 @@ class RetrievalQAChain_Chains implements INode {
                 if (shouldStreamResponse) {
                     streamResponse(sseStreamer, chatId, e.message)
                 }
-                // streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
