+++ packages/components/nodes/chains/VectaraChain/VectaraChain.ts
@@ -272,7 +272,6 @@ class VectaraChain_Chains implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                //streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
