+++ packages/components/nodes/chains/VectorDBQAChain/VectorDBQAChain.ts
@@ -77,7 +77,6 @@ class VectorDBQAChain_Chains implements INode {
                 // if (options.shouldStreamResponse) {
                 //     streamResponse(options.sseStreamer, options.chatId, e.message)
                 // }
-                // streamResponse(options.socketIO && options.socketIOClientId, e.message, options.socketIO, options.socketIOClientId)
                 return formatResponse(e.message)
             }
         }
