+++ packages/server/src/controllers/predictions/index.ts
@@ -67,23 +67,7 @@ const createPrediction = async (req: Request, res: Response, next: NextFunction)
             const apiResponse = await predictionsServices.buildChatflow(req)
             if (streamable?.isStreaming && isStreamingRequested) {
                 const sseStreamer = getRunningExpressApp().sseStreamer
-                const metadataJson: any = {}
-                if (apiResponse.chatId) {
-                    metadataJson['chatId'] = apiResponse.chatId
-                }
-                if (apiResponse.chatMessageId) {
-                    metadataJson['chatMessageId'] = apiResponse.chatMessageId
-                }
-                if (apiResponse.question) {
-                    metadataJson['question'] = apiResponse.question
-                }
-                if (apiResponse.sessionId) {
-                    metadataJson['sessionId'] = apiResponse.sessionId
-                }
-                if (apiResponse.memoryType) {
-                    metadataJson['memoryType'] = apiResponse.memoryType
-                }
-                sseStreamer.streamCustomEvent(apiResponse.chatId, 'metadata', metadataJson)
+                sseStreamer.streamMetadataEvent(apiResponse.chatId, apiResponse)
                 sseStreamer.removeClient(apiResponse.chatId)
                 return
             }
